task: clean project root
setup.sh appears to be a developer utility to assist in bootstrapping a debian
based system.  It is intended to help install the dependencies needed to run
node, and not used in any automated deploy pipeline.

Instructions have been updated to suite this change which simply moves the
developer tool out of the project repo and into the bin subdirectory.